Move operators to their own package
